mushroom/util/util.py
import yaml
from mushroom.exception import MushroomException
import os
import sys

def read_yaml_file(file_path:str)->dict:
    """
    Reads a YAML file and returns the contents as a dictionary.
    file_path: str
    """
    try:
        with open(file_path, "rb") as yaml_file:
            return yaml.safe_load(yaml_file)
    except Exception as e:
        raise MushroomException(e,sys) from e