requirements.txt
numpy
Flask
gunicorn
pandas
numpy
scikit-learn
ipykernel
PyYAML
-e .